requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions
binance-connector
azure-data-tables
binance-futures-connector
azure-storage-queue
pandas
pandas-ta
ta
numpy<2.0
setuptools